+++ backend/tests/baserow/contrib/database/api/views/form/test_form_view_views.py
@@ -34,6 +34,7 @@ def test_create_form_view(api_client, data_fixture):
     assert response_json["description"] == ""
     assert response_json["cover_image"] is None
     assert response_json["logo_image"] is None
+    assert response_json["submit_text"] == "Submit"
     assert response_json["submit_action"] == "MESSAGE"
     assert response_json["submit_action_redirect_url"] == ""
 
@@ -48,6 +49,7 @@ def test_create_form_view(api_client, data_fixture):
     assert form.description == ""
     assert form.cover_image is None
     assert form.logo_image is None
+    assert form.submit_text == "Submit"
     assert form.submit_action == "MESSAGE"
     assert form.submit_action_redirect_url == ""
     assert "filters" not in response_json
@@ -64,6 +66,7 @@ def test_create_form_view(api_client, data_fixture):
             "description": "Description",
             "cover_image": {"name": user_file_1.name},
             "logo_image": {"name": user_file_2.name},
+            "submit_text": "NEW SUBMIT",
             "submit_action": "REDIRECT",
             "submit_action_redirect_url": "https://localhost",
         },
@@ -80,6 +83,7 @@ def test_create_form_view(api_client, data_fixture):
     assert response_json["description"] == "Description"
     assert response_json["cover_image"]["name"] == user_file_1.name
     assert response_json["logo_image"]["name"] == user_file_2.name
+    assert response_json["submit_text"] == "NEW SUBMIT"
     assert response_json["submit_action"] == "REDIRECT"
     assert response_json["submit_action_redirect_url"] == "https://localhost"
 
@@ -92,6 +96,7 @@ def test_create_form_view(api_client, data_fixture):
     assert form.description == "Description"
     assert form.cover_image_id == user_file_1.id
     assert form.logo_image_id == user_file_2.id
+    assert form.submit_text == "NEW SUBMIT"
     assert form.submit_action == "REDIRECT"
     assert form.submit_action_redirect_url == "https://localhost"
 
@@ -129,6 +134,7 @@ def test_update_form_view(api_client, data_fixture):
             "type": "form",
             "public": True,
             "title": "Title",
+            "submit_text": "Patched Submit",
             "description": "Description",
             "cover_image": {"name": user_file_1.name},
             "logo_image": {"name": user_file_2.name},
@@ -148,6 +154,7 @@ def test_update_form_view(api_client, data_fixture):
     assert response_json["description"] == "Description"
     assert response_json["cover_image"]["name"] == user_file_1.name
     assert response_json["logo_image"]["name"] == user_file_2.name
+    assert response_json["submit_text"] == "Patched Submit"
     assert response_json["submit_action"] == "REDIRECT"
     assert response_json["submit_action_redirect_url"] == "https://localhost"
 
@@ -159,6 +166,7 @@ def test_update_form_view(api_client, data_fixture):
     assert form.description == "Description"
     assert form.cover_image_id == user_file_1.id
     assert form.logo_image_id == user_file_2.id
+    assert form.submit_text == "Patched Submit"
     assert form.submit_action == "REDIRECT"
     assert form.submit_action_redirect_url == "https://localhost"