+++ changelog.md
@@ -15,6 +15,7 @@
   `/database/views/${viewId}/rotate-slug/`.
 * Increased maximum length of application name to 160 characters.
 * Fixed copying/pasting for date field.
+* Allow changing the text of the submit button in the form view.
 
 ## Released (2021-11-25)